Login and ssh key push.

+++ docker/proxy/entrypoint.sh
@@ -1,12 +1,22 @@
 #!/bin/sh -x
 
+MD5_CURR=""
+MD5_PREV=""
+
 configure() {
-    if [ ! -f "${PID}" ] || ! kill -0 $(cat ${PID}); then
+    envsubst < ${CONF}.tmpl > ${CONF}
+    MD5_CURR=$(md5sum ${CONF})
+
+    if [ -f ${PID} ] && kill -0 $(cat ${PID}); then
+        # It's already running, did config change?
+        if [ "${MD5_CURR}" != "${MD5_PREV}" ]; then
+            kill -HUP $(cat ${PID})
+        fi
+    else
         nginx -g 'daemon off;' &
         echo -n "$!" > ${PID}
-    else
-        kill -HUP $(cat ${PID})
     fi
+    MD5_PREV=${MD5_CURR}
 }
 
 trap 'configure' 1
